refactor: implement modular architecture for env-help system
Major refactoring addressing latest Copilot feedback:

- Move 260+ line ENV_CATEGORIES dictionary to separate env_help_config module
- Extract truncate_text utility function to utils.text module
- Pre-compute ENV_HELP_CATEGORY_CHOICES to avoid runtime list generation
- Reduce main.py complexity from 1400+ to ~1100 lines
- Fix import organization to comply with PEP8 standards

All functionality preserved:
- 16/16 CLI tests pass
- Environment variable help system fully operational
- Modular architecture for better maintainability

Github-Issue:#34
